+++ scripts/dual_sync.sh
@@ -88,7 +88,7 @@ sync_module "files"
 sync_module "fuzzy"
 sync_module "git"
 sync_module "hipatterns"
-sync_module "hues" colors/randomhue.lua
+sync_module "hues" colors/miniwinter.lua colors/minispring.lua colors/minisummer.lua colors/miniautumn.lua colors/randomhue.lua
 sync_module "icons"
 sync_module "indentscope"
 sync_module "jump"