+++ tests/test_extra.lua
@@ -945,21 +945,21 @@ T['pickers']['colorschemes()']['works'] = function()
   validate_picker_name('Colorschemes')
   type_keys('^mini')
 
-  -- Should find minicyan and minischeme
+  -- Should find bundled 'mini*' color schemes
   child.expect_screenshot({ ignore_text = { 14 } })
 
   -- Should have proper preview
   type_keys('<Tab>')
-  eq(child.g.colors_name, 'minicyan')
+  eq(child.g.colors_name, 'miniautumn')
   child.expect_screenshot({ ignore_text = { 14 } })
 
   -- Should properly choose
   type_keys('<CR>')
-  eq(child.g.colors_name, 'minicyan')
-  expect.match(child.cmd_capture('hi Normal'), 'guibg=#0a2a2a')
+  eq(child.g.colors_name, 'miniautumn')
+  expect.match(child.cmd_capture('hi Normal'), 'guibg=#262029')
 
   -- Should return chosen value
-  eq(child.lua_get('_G.return_item'), 'minicyan')
+  eq(child.lua_get('_G.return_item'), 'miniautumn')
 end
 
 T['pickers']['colorschemes()']['works with preview'] = function()
@@ -974,13 +974,13 @@ T['pickers']['colorschemes()']['works with preview'] = function()
   end
 
   type_keys('<Tab>')
-  validate('minicyan', '#0a2a2a')
+  validate('miniautumn', '#262029')
 
   type_keys('<C-n>')
-  validate('minischeme', '#112641')
+  validate('minicyan', '#0a2a2a')
 
   type_keys('<C-p>')
-  validate('minicyan', '#0a2a2a')
+  validate('miniautumn', '#262029')
 end
 
 T['pickers']['colorschemes()']['previews with original background'] = function()
@@ -988,7 +988,7 @@ T['pickers']['colorschemes()']['previews with original background'] = function()
 
   child.o.background = 'dark'
   pick_colorschemes()
-  type_keys('^mini', '<C-n>', '<Tab>')
+  type_keys('^mini', '<C-n>', '<C-n>', '<Tab>')
   eq(child.o.background, 'light')
   eq(child.g.colors_name, 'miniforcebg')
 
@@ -1001,7 +1001,7 @@ T['pickers']['colorschemes()']['can choose marked'] = function()
   pick_colorschemes()
   type_keys('^mini')
   type_keys('<C-x>', '<C-n>', '<C-x>', '<M-CR>')
-  eq(child.g.colors_name, 'minicyan')
+  eq(child.g.colors_name, 'miniautumn')
 end
 
 T['pickers']['colorschemes()']["can cancel with 'mini.colors'"] = function()
@@ -1012,7 +1012,7 @@ T['pickers']['colorschemes()']["can cancel with 'mini.colors'"] = function()
 
   child.lua_notify('_G.return_item = MiniExtra.pickers.colorschemes()')
   type_keys('^mini', '<Tab>')
-  eq(child.g.colors_name, 'minicyan')
+  eq(child.g.colors_name, 'miniautumn')
 
   -- Should trigger 'ColorScheme' event
   child.cmd('au ColorScheme * lua _G.n = (_G.n or 0) + 1')
@@ -1035,21 +1035,21 @@ T['pickers']['colorschemes()']["can cancel without 'mini.colors'"] = function()
     end
   ]])
 
-  child.cmd('colorscheme minischeme')
+  child.cmd('colorscheme miniwinter')
   -- These customizations can not persist even if there was preview
   child.api.nvim_set_hl(0, 'Normal', { fg = '#000000' })
   child.g.terminal_color_0 = '#010101'
 
   child.lua_notify('_G.return_item = MiniExtra.pickers.colorschemes()')
   type_keys('^mini', '<Tab>')
-  eq(child.g.colors_name, 'minicyan')
+  eq(child.g.colors_name, 'miniautumn')
 
   type_keys('<C-c>')
 
   eq(child.lua_get('_G.return_item'), vim.NIL)
-  eq(child.g.colors_name, 'minischeme')
-  expect.match(child.cmd_capture('hi Normal'), 'guifg=#e2e98f')
-  eq(child.g.terminal_color_0, '#112641')
+  eq(child.g.colors_name, 'miniwinter')
+  expect.match(child.cmd_capture('hi Normal'), 'guifg=#d8d4cd')
+  eq(child.g.terminal_color_0, '#000f15')
 
   -- Should work if there is no `g:colors_name` defined before starting picker
   child.g.colors_name = nil
@@ -1066,10 +1066,10 @@ T['pickers']['colorschemes()']['restores original color scheme only if needed'] 
 end
 
 T['pickers']['colorschemes()']['respects `local_opts.names`'] = function()
-  pick_colorschemes({ names = { 'randomhue', 'minischeme' } })
-  eq(get_picker_items(), { 'randomhue', 'minischeme' })
+  pick_colorschemes({ names = { 'minisummer', 'minispring' } })
+  eq(get_picker_items(), { 'minisummer', 'minispring' })
   type_keys('<Tab>')
-  eq(child.g.colors_name, 'randomhue')
+  eq(child.g.colors_name, 'minisummer')
   type_keys('<C-c>')
 
   -- Should validate